connected test/j0 to convex
